+++ hypothesis-python/tests/quality/test_float_shrinking.py
@@ -8,17 +8,12 @@
 # v. 2.0. If a copy of the MPL was not distributed with this file, You can
 # obtain one at https://mozilla.org/MPL/2.0/.
 
-import math
-import struct
-
 import pytest
 
-from hypothesis import example, given, seed, strategies as st
+from hypothesis import example, given, strategies as st
 from hypothesis.internal.compat import ceil
-from hypothesis.internal.conjecture.data import ConjectureData
 
 from tests.common.debug import minimal
-from tests.conjecture.common import shrinking_from
 
 
 def test_shrinks_to_simple_floats():
@@ -50,22 +45,3 @@ def test_shrinks_downwards_to_integers_when_fractional(b):
         ).filter(lambda x: int(x) != x)
     )
     assert g == b + 0.5
-
-
-@pytest.mark.parametrize("nan",
-    [
-        math.nan,
-        -math.nan,
-        struct.unpack('d', struct.pack('Q', 0xfff8000000000001))[0]
-    ]
-)
-def test_shrinks_to_canonical_nan(nan):
-    @shrinking_from([nan])
-    def shrinker(data: ConjectureData):
-        value = data.draw_float()
-        if math.isnan(value):
-            data.mark_interesting()
-
-    shrinker.shrink()
-    assert len(shrinker.choices) == 1
-    assert struct.pack("d", shrinker.choices[0]) == struct.pack("d", math.nan)
